# app/backend/docs/swagger.yaml
basePath: /
definitions:
  types.CreateVideoRequest:
    properties:
      description:
        example: This is a sample video description
        type: string
      s3_key:
        example: videos/sample-video.mp4
        type: string
      title:
        example: Sample Video
        type: string
    required:
    - s3_key
    - title
    type: object
  types.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  types.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
  types.VideoResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: This is a sample video description
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      s3_key:
        example: videos/sample-video.mp4
        type: string
      title:
        example: Sample Video
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for Video Annotation Platform - Phase 1 Foundation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Video Annotation Platform API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Returns the health status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponse'
      summary: Health check endpoint
      tags:
      - health
  /videos:
    get:
      consumes:
      - application/json
      description: Retrieves all video metadata records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.VideoResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all videos
      tags:
      - videos
    post:
      consumes:
      - application/json
      description: Creates a new video metadata record
      parameters:
      - description: Video metadata
        in: body
        name: video
        required: true
        schema:
          $ref: '#/definitions/types.CreateVideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.VideoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create video metadata
      tags:
      - videos
  /videos/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a video metadata by its ID
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.VideoResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get video by ID
      tags:
      - videos
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"